% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregion.band.R
\name{fregion.band}
\alias{fregion.band}
\title{Make confidence bands}
\usage{
fregion.band(x, cov, N = 1, type = c("Bs", "BEc"), conf.level = c(0.95),
  grid.size = 200, inv.method = "gamma.approx", prec = NULL,
  sim.size = 10000)
}
\arguments{
\item{x}{Observed function. It can be either a vector, of fd object in 'fda' package.}

\item{cov}{The covaraicne operator (as matrix) for the random function x.}

\item{N}{Sample size. It should be '1' if 'cov' is the covariance operator for 'x' itself. If 'x' is the mean of an i.i.d. sample of size 'N' and the 'cov' is the covariance operator of each unit in the sample, then this 'N' can be used here.}

\item{type}{This specifies which bands to be constructed. It should be a collection of the followings : "Bs", "BEc", "BEc1".}

\item{conf.level}{A vector of confidence levels for the bands to achieve.}

\item{grid.size}{This determines on how fine grid the bands will be constructed. Note that this parameter is used only when 'x' is fd object and 'cov' is bifd object.}

\item{inv.method}{(Currently Not Used) This determines how the inverse of the sum of chi squares distribution will be achieved. It can be either "gamma.approx", or "inv.imhof".}

\item{prec}{(Currently Not Used) This determines the accuracy of imhof. It's used only when inv.method is inv.imhof.}

\item{sim.size}{This determines the simulation size for Bs}
}
\value{
bands returns bands
}
\description{
Make confidence bands
}
\examples{
p <- 100 ; N <- 50
grid <- make.grid(p)
mu <- mean.f.poly(grid,c(0,1)) ; names(mu) <- grid
cov.m <- make.cov.m(cov.f=cov.f.st.matern, grid=grid, cov.f.params=c(2/3,1,1/4))
X <- make.sample(mu,cov.m,N)
hat.mu <- rowMeans(X)
hat.cov.m <- crossprod(X - hat.mu) / (N-1)
fregion.band(hat.mu, hat.cov.m, N, type=c("Bs","BEc"), conf.level=c(0.90,0.95))
}

