% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregion.rect.R
\name{fregion.rect}
\alias{fregion.rect}
\title{Make Rectangular Confidence Regoion}
\usage{
fregion.rect(x, cov, N = 1, type = "Rz", conf.level = 0.95,
  pc.cut = 0.999, df = NULL)
}
\arguments{
\item{x}{Observed function. It can be either a vector or fd object in 'fda' package.}

\item{cov}{The covaraicne operator (as matrix, or 'bifd' object) for the random function x.}

\item{N}{Sample size. It should be '1' if 'cov' is the covariance operator for 'x' itself. If 'x' is the mean of an i.i.d. sample of size 'N' and the 'cov' is the covariance operator of each unit in the sample, then this 'N' can be used here.}

\item{type}{This specifies which rectangular region to be constructed. It should be either "Rz", "Rzs", "Rz1", "Rz1s".}

\item{conf.level}{This specifies confidence level.}

\item{pc.cut}{This specifies number of fpc to use. For integer values, fpc up to those values will be used. If it's a value from 0 to 1, this specifies the proportion of variance that should be explained by the fpcs. If it is 0, all the available fpcs will be used as long as the size of eigenvalues are greater than .Machine$double.eps.}

\item{df}{Degree of freedom used for "Rzs" and "Rz1s". If null, N-1 will be used.}
}
\description{
Make Rectangular Confidence Regoion
}
\examples{
p <- 100 ; N <- 50
grid <- make.grid(p)
mu <- mean.f.poly(grid,c(0,1)) ; names(mu) <- grid
cov.m <- make.cov.m(cov.f <- cov.f.st.matern, grid=grid, cov.f.params=c(2/3,1,1/4))
X <- make.sample(mu,cov.m,N)
hat.mu <- rowMeans(X)
hat.cov.m <- crossprod(X - hat.mu) / (N-1)
fregion.rect(hat.mu, hat.cov.m, N, type="Rz", conf.level= 0.95)
}

