% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fregion.band.R
\name{plot.fregion.band}
\alias{plot.fregion.band}
\title{Visualize confidence bands}
\usage{
\method{plot}{fregion.band}(band, col = NULL, lty = NULL, center = TRUE)
}
\arguments{
\item{band}{A 'fregion.band' object, the output from 'region.band()' funciton.}

\item{col}{a vector of colors to be used for the bands.}

\item{lty}{a vector of colors to be used for the bands.}

\item{center}{Whether to include the original estimate.}
}
\description{
Visualize confidence bands
}
\examples{
plot(band)
}

